avoir circular import when adding schema feature
