+++ dcm2bids/utils/tools.py
@@ -9,9 +9,9 @@ from pathlib import Path
 from urllib import error, request
 from subprocess import getoutput
 from shutil import which
+
 from dcm2bids.version import __version__
 from dcm2bids.utils.io import load_json, save_json
-from dcm2bids.utils.utils import _version_newer
 
 logger = logging.getLogger(__name__)
 
@@ -52,6 +52,61 @@ def _save_version_cache(cache: dict, log_dir: str | Path) -> None:
         save_json(filename=path, data=cache)
     except Exception:
         logger.debug("Failed to write version cache; ignoring.", exc_info=True)
+
+def _normalize_version(value):
+    """
+    Normalize a version string into a tuple of integer parts, when possible.
+
+    - Strips a leading 'v' (e.g. 'v1.11.0' -> '1.11.0')
+    - Splits on '.'
+    - Converts each part to int if possible
+    - Returns the original value if it cannot be parsed in a simple numeric form
+    """
+    if not isinstance(value, str):
+        return value
+
+    s = value.strip()
+    if not s:
+        return value
+
+    # BIDS tag style: v1.2.3
+    s = s.lstrip("v")
+
+    parts = s.split(".")
+    normalized = []
+    for p in parts:
+        p = p.strip()
+        if not p:
+            # Empty segment; treat as zero
+            normalized.append(0)
+            continue
+        try:
+            normalized.append(int(p))
+        except ValueError:
+            # If any part is not numeric, bail return the original string
+            return value
+
+    return tuple(normalized)
+
+def _version_newer(latest, current):
+    """
+    Compare two versions, preferring normalized numeric comparison
+    and falling back to string comparison.
+    
+    Returns True if `latest` is considered newer than `current`.
+    """
+    norm_latest = _normalize_version(latest)
+    norm_current = _normalize_version(current)
+
+    if isinstance(norm_latest, tuple) and isinstance(norm_current, tuple):
+        # Padding to compare same length
+        max_len = max(len(norm_latest), len(norm_current))
+        norm_latest += (0,) * (max_len - len(norm_latest))
+        norm_current += (0,) * (max_len - len(norm_current))
+        return norm_latest > norm_current
+
+    # Fallback: string comparison as before
+    return str(latest) > str(current)
 
 
 def is_tool(name):
